Dark-Web-Hunter-main/requirements.txt
tk
Pillow
requests
beautifulsoup4
